Factor database code out to new file
Just improving encapsulation and readability
